refactor(factory): keep the app shell in factory instead of the design system
