+++ mastracode/factory-ui/src/ui/layouts/__tests__/AppShell.msw.test.tsx
@@ -1,8 +1,7 @@
-// @vitest-environment jsdom
 import { cleanup, render, screen } from '@testing-library/react';
 import { afterEach, describe, expect, it } from 'vitest';
 
-import { AppShell } from './app-shell';
+import { AppShell } from '../AppShell';
 
 afterEach(() => {
   cleanup();
