--- packages/playground-ui/src/ds/components/AppShell/index.ts
@@ -1,2 +1,0 @@
-export { AppShell } from './app-shell';
-export type { AppShellProps } from './app-shell';
